a2g2_mc_pip_requirements.txt
Django
django-extensions
django-taggit
djangorestframework
psutil
requests
django-filter
pexpect
jinja2
pyyaml
jsonfield
